Mint

+++ Codigos/Primitivas/Modular-Int/mint.cpp
@@ -12,7 +12,7 @@ template <int mod> struct Mint {
         if (v < 0) {
             v += mod;
         }
-        val = (int)v;
+        val = int(v);
     }
     m pwr(m b, ll e) {
         m res = 1;
@@ -36,7 +36,7 @@ template <int mod> struct Mint {
         return *this;
     }
     m &operator*=(const m &o) {
-        val = (ll)val * o.val % mod;
+        val = int((ll)val * o.val % mod);
         return *this;
     }
     m &operator/=(const m &o) { return *this *= pwr(o, mod - 2); }
